add sos information
